requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
spotipy==2.19.0
urllib3==1.26.9
